+++ src/Entity/Season.php
@@ -33,17 +33,13 @@ class Season
     #[ORM\Column]
     private ?bool $isActive = false;
 
-    /**
-     * @var Collection<int, WorldChampion>
-     */
-    #[ORM\OneToMany(targetEntity: WorldChampion::class, mappedBy: 'season')]
-    private Collection $worldChampions;
+    #[ORM\ManyToOne(inversedBy: 'worldChampionSeasons')]
+    private ?Driver $worldChampion = null;
 
     public function __construct()
     {
         $this->worldChampionBets = new ArrayCollection();
         $this->races = new ArrayCollection();
-        $this->worldChampions = new ArrayCollection();
     }
 
     public function getId(): ?int
@@ -135,32 +131,14 @@ class Season
         return $this;
     }
 
-    /**
-     * @return Collection<int, WorldChampion>
-     */
-    public function getWorldChampions(): Collection
+    public function getWorldChampion(): ?Driver
     {
-        return $this->worldChampions;
+        return $this->worldChampion;
     }
 
-    public function addWorldChampion(WorldChampion $worldChampion): static
+    public function setWorldChampion(?Driver $worldChampion): static
     {
-        if (!$this->worldChampions->contains($worldChampion)) {
-            $this->worldChampions->add($worldChampion);
-            $worldChampion->setSeason($this);
-        }
-
-        return $this;
-    }
-
-    public function removeWorldChampion(WorldChampion $worldChampion): static
-    {
-        if ($this->worldChampions->removeElement($worldChampion)) {
-            // set the owning side to null (unless already changed)
-            if ($worldChampion->getSeason() === $this) {
-                $worldChampion->setSeason(null);
-            }
-        }
+        $this->worldChampion = $worldChampion;
 
         return $this;
     }